WIP: fixing up derive
